pyproject.toml
[project]
name = "lead-qualifier"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mcp[cli]>=1.27.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.45.0",
]
